+++ src/blueprints/telegram_bot/webhook/commands/help.py
@@ -69,10 +69,10 @@ def handle(*args, **kwargs):
         "Folder name should starts from root, "
         f'nested folders should be separated with "{to_code("/")}" character.'
         "\n"
-        f"{CommandName.SPACE.value} — get information about remaining space."
-        "\n"
         f"{CommandName.ELEMENT_INFO.value} — get information about file or folder. "
         "Send full path of element with this command."
+        "\n"
+        f"{CommandName.SPACE_INFO.value} — get information about remaining space."
         "\n"
         f"{CommandName.DISK_INFO.value} — get information about your Yandex.Disk. "
         "\n\n"